requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
sentence-transformers==3.0.1
faiss-cpu==1.12.0
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyPDF2==3.0.1
python-docx==0.8.11
